# pulsar-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}/control/delay:
    post:
      tags:
        - control
      summary: Delay execution
      operationId: delay
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelayRequest'
      responses:
        '200':
          description: Delay completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/control/pause:
    post:
      tags:
        - control
      summary: Pause session execution
      operationId: pause
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/control/stop:
    post:
      tags:
        - control
      summary: Stop session execution
      operationId: stop
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/agent/run:
    post:
      tags:
        - agent
      summary: Run an autonomous agent task
      description: |
        Runs an autonomous loop (observe -> act -> ...) attempting to fulfill
        the user goal described in the request.
      operationId: agentRun
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Agent task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/agent/observe:
    post:
      tags:
        - agent
      summary: Observe the page
      description: |
        Observes the page given an instruction and returns potential actions
        that could be taken.
      operationId: agentObserve
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentObserveRequest'
      responses:
        '200':
          description: Observation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentObserveResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/agent/act:
    post:
      tags:
        - agent
      summary: Execute a single action
      description: |
        Executes a single observe->act cycle for the supplied action.
      operationId: agentAct
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentActRequest'
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentActResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/agent/extract:
    post:
      tags:
        - agent
      summary: Extract structured data from the page
      description: |
        Extracts structured data from the page using an instruction and
        optional JSON schema.
      operationId: agentExtract
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExtractRequest'
      responses:
        '200':
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentExtractResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/agent/summarize:
    post:
      tags:
        - agent
      summary: Summarize page content
      description: |
        Summarizes text captured from the current page or a specific element.
      operationId: agentSummarize
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentSummarizeRequest'
      responses:
        '200':
          description: Summary result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSummarizeResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/agent/clearHistory:
    post:
      tags:
        - agent
      summary: Clear agent history
      description: |
        Clears the agent's history so new tasks remain unaffected by previous ones.
      operationId: agentClearHistory
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: History cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentClearHistoryResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== PulsarSession Endpoints ==========
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
...
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
...
    DelayRequest:
      type: object
      required:
        - ms
      properties:
        ms:
          type: integer
          minimum: 0
          description: Delay duration in milliseconds
...
    AgentRunRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          description: The user goal or instruction to fulfill
        multiAct:
          type: boolean
          default: false
          description: Whether each act forms a new chained context
        modelName:
          type: string
          description: Optional LLM model name to use
        variables:
          type: object
          additionalProperties:
            type: string
          description: Extra variables for prompt/tool
        domSettleTimeoutMs:
          type: integer
          format: int64
          description: Timeout in ms to wait for DOM settling
        timeoutMs:
          type: integer
          format: int64
          description: Overall timeout in ms for this action
    AgentRunResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            success:
              type: boolean
            message:
              type: string
            historySize:
              type: integer
            processTraceSize:
              type: integer
    AgentObserveRequest:
      type: object
      properties:
        instruction:
          type: string
          description: The observation instruction
        modelName:
          type: string
          description: Optional LLM model name
        domSettleTimeoutMs:
          type: integer
          format: int64
        returnAction:
          type: boolean
          description: Whether to return actionable tool calls
        drawOverlay:
          type: boolean
          default: true
          description: Whether to highlight interactive elements
    AgentObserveResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/ObserveResult'
    ObserveResult:
      type: object
      properties:
        locator:
          type: string
          description: DOM node locator (frameIndex,backendNodeId)
        domain:
          type: string
          description: Tool call domain (driver, browser, etc.)
        method:
          type: string
          description: Tool call method (click, type, etc.)
        arguments:
          type: object
          additionalProperties: true
          description: Tool call arguments
        description:
          type: string
        screenshotContentSummary:
          type: string
        currentPageContentSummary:
          type: string
        nextGoal:
          type: string
        thinking:
          type: string
        summary:
          type: string
        keyFindings:
          type: string
        nextSuggestions:
          type: array
          items:
            type: string
    AgentActRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          description: The action command to execute
        multiAct:
          type: boolean
          default: false
        modelName:
          type: string
        variables:
          type: object
          additionalProperties:
            type: string
        domSettleTimeoutMs:
          type: integer
          format: int64
        timeoutMs:
          type: integer
          format: int64
    AgentActResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            success:
              type: boolean
            message:
              type: string
            action:
              type: string
            isComplete:
              type: boolean
            expression:
              type: string
              description: The tool expression with parameters
    AgentExtractRequest:
      type: object
      required:
        - instruction
      properties:
        instruction:
          type: string
          description: The extraction instruction
        schema:
          $ref: '#/components/schemas/ExtractionSchema'
        modelName:
          type: string
        domSettleTimeoutMs:
          type: integer
          format: int64
        selector:
          type: string
          description: Optional CSS selector to scope extraction
    ExtractionSchema:
      type: object
      properties:
        type:
          type: string
          default: object
        properties:
          type: object
          additionalProperties: true
        required:
          type: array
          items:
            type: string
    AgentExtractResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            success:
              type: boolean
            message:
              type: string
            data:
              type: object
              additionalProperties: true
    AgentSummarizeRequest:
      type: object
      properties:
        instruction:
          type: string
          description: Optional guidance for summarization
        selector:
          type: string
          description: Optional CSS selector to limit which element to summarize
    AgentSummarizeResponse:
      type: object
      properties:
        value:
          type: string
          description: The generated summary text
    AgentClearHistoryResponse:
      type: object
      properties:
        value:
          type: boolean
          description: Whether history was cleared successfully

    # ========== PulsarSession Schemas ==========
...